Moved stories to separate project (and folder)
