+++ package.json
@@ -29,17 +29,17 @@
         "node": ">=10"
     },
     "scripts": {
-        "start": "tsdx watch --tsconfig tsconfig.build.json",
-        "build": "tsdx build --tsconfig tsconfig.build.json",
+        "start": "tsdx watch",
+        "build": "tsdx build",
         "test": "jest",
-        "lint": "tsdx lint --tsconfig tsconfig.build.json",
-        "prepare": "tsdx build --tsconfig tsconfig.build.json",
+        "lint": "tsdx lint",
+        "prepare": "tsdx build",
         "size": "size-limit",
         "postinstall": "husky install",
         "prepack": "pinst --disable",
         "postpack": "pinst --enable",
-        "storybook": "start-storybook -p 6006",
-        "build-storybook": "build-storybook"
+        "storybook": "npm --prefix stories run storybook -p 6006",
+        "build-storybook": "npm --prefix stories run build-storybook"
     },
     "peerDependencies": {
         "solid-js": ">=1.0.0<2"
@@ -48,25 +48,11 @@
         "@babel/core": "^7.18.6",
         "@jest/types": "^28.1.3",
         "@size-limit/preset-small-lib": "^7.0.8",
-        "@storybook/addon-actions": "^6.5.9",
-        "@storybook/addon-essentials": "^6.5.9",
-        "@storybook/addon-interactions": "^6.5.9",
-        "@storybook/addon-links": "^6.5.9",
-        "@storybook/addon-postcss": "2.0.0",
-        "@storybook/builder-webpack4": "^6.5.9",
-        "@storybook/html": "^6.5.9",
-        "@storybook/manager-webpack4": "^6.5.9",
-        "@storybook/node-logger": "^6.5.9",
-        "@storybook/testing-library": "^0.0.13",
-        "autoprefixer": "^10.4.7",
-        "babel-loader": "^8.2.5",
         "babel-preset-solid": "^1.4.6",
         "husky": "^8.0.1",
         "jest": "^28.1.3",
         "pinst": "^3.0.0",
-        "postcss": "^8.4.14",
         "size-limit": "^7.0.8",
-        "tailwindcss": "^3.1.6",
         "ts-jest": "^28.0.6",
         "tsdx": "^0.14.1",
         "tslib": "^2.4.0",
@@ -78,7 +64,6 @@
     "resolutions": {
         "**/typescript": "^4.1.0",
         "**/ts-jest": "^28.0.6",
-        "**/jest": "^28.1.3",
-        "**/postcss": "^8.0.0"
+        "**/jest": "^28.1.3"
     }
 }
