+++ stories/.storybook/.babelrc
@@ -1,0 +1,3 @@
+{
+    "presets": ["solid", "@babel/preset-typescript"]
+}
