+++ stories/.storybook/main.js
@@ -1,11 +1,10 @@
 module.exports = {
-    stories: ['../stories/**/*.stories.mdx', '../stories/**/*.stories.@(js|jsx|ts|tsx)'],
+    stories: ['../src/*.stories.mdx', '../src/*.stories.@(js|jsx|ts|tsx)'],
     addons: [
         '@storybook/addon-links',
         '@storybook/addon-essentials',
         '@storybook/addon-interactions',
-        '@storybook/addon-postcss',
-        '@storybook/node-logger'
+        '@storybook/addon-postcss'
     ],
     framework: '@storybook/html'
 };
