+++ stories/package.json
@@ -1,0 +1,61 @@
+{
+    "name": "solidjs-swipe-card-storybook",
+    "description": "A SolidJS swipeable card component (tinder-like) demo",
+    "author": "mardisen",
+    "repository": "mardisen/solidjs-swipe-card",
+    "version": "0.1.0",
+    "license": "MIT",
+    "private": false,
+    "keywords": [
+        "solid-js",
+        "solid",
+        "tinder",
+        "card",
+        "web",
+        "swipeable",
+        "swipe",
+        "physics",
+        "deck",
+        "drag",
+        "draggable"
+    ],
+    "files": [
+        "src"
+    ],
+    "engines": {
+        "node": ">=10"
+    },
+    "scripts": {
+        "storybook": "start-storybook -p 6006",
+        "build-storybook": "build-storybook"
+    },
+    "peerDependencies": {
+        "solid-js": ">=1.0.0<2"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.18.6",
+        "@babel/preset-typescript": "^7.18.6",
+        "@storybook/addon-actions": "^6.5.9",
+        "@storybook/addon-essentials": "^6.5.9",
+        "@storybook/addon-interactions": "^6.5.9",
+        "@storybook/addon-links": "^6.5.9",
+        "@storybook/addon-postcss": "2.0.0",
+        "@storybook/builder-webpack4": "^6.5.9",
+        "@storybook/html": "^6.5.9",
+        "@storybook/manager-webpack4": "^6.5.9",
+        "@storybook/testing-library": "^0.0.13",
+        "autoprefixer": "^10.4.7",
+        "babel-loader": "^8.2.5",
+        "babel-preset-solid": "^1.4.6",
+        "postcss": "^8.4.14",
+        "tailwindcss": "^3.1.6",
+        "tslib": "^2.4.0",
+        "typescript": "^4.7.4"
+    },
+    "dependencies": {
+        "solid-js": "^1.4.7"
+    },
+    "resolutions": {
+        "**/postcss": "^8.0.0"
+    }
+}
