+++ stories/tsconfig.json
@@ -1,6 +1,6 @@
 {
     // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-    "include": ["src", "types"],
+    "include": ["src"],
     "compilerOptions": {
         "module": "esnext",
         "lib": ["dom", "esnext"],
@@ -10,7 +10,7 @@
         // output .js.map sourcemap files for consumers
         "sourceMap": true,
         // match output dir to input dir. e.g. dist/index instead of dist/src/index
-        "rootDir": "./src",
+        "rootDirs": ["./src", "../../src"],
         // stricter type-checking for stronger correctness. Recommended by TS
         "strict": false,
         // linter checks for common issues
